Added CloudExtactionsController && Moved methods from CR Controller to CE Controller.
